+++ src/database/MappedDatabaseHandler.java
@@ -1,0 +1,17 @@
+package database;
+
+import networking.messageHandlers.MappedMessageHandler;
+import networking.messageHandlers.MessageHandler;
+import networking.messages.Message;
+
+import java.util.TreeMap;
+
+/**
+ * Created by Nikolay on 1/7/2017.
+ */
+public class MappedDatabaseHandler extends MappedMessageHandler {
+
+    public MappedDatabaseHandler(){
+        super();
+    }
+}
