+++ src/database/databaseController/DatabaseAuthorizationController.java
@@ -1,8 +1,9 @@
-package database;
+package database.databaseController;
 
 import dataModel.models.Client;
 import dataModel.models.SystemProfile;
 import dataModel.models.SystemProfileType;
+import dataModel.PasswordConver;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
