+++ src/database/databaseController/DatabaseSystemProfileController.java
@@ -1,4 +1,4 @@
-package database;
+package database.databaseController;
 
 import dataModel.models.SystemProfileType;
