+++ src/networking/messages/response/BalanceResponse.java
@@ -1,0 +1,36 @@
+package networking.messages.response;
+
+import dataModel.Money;
+
+import java.sql.ResultSet;
+
+/**
+ * Created by Nikolay on 1/8/2017.
+ */
+public class BalanceResponse extends Response {
+
+    public static final String TYPE = "BALANCERESPONSE";
+
+    private String userName;
+
+    private Money amount;
+
+    public BalanceResponse(String userName, Money money ) {
+        super(this.TYPE, "Data for "+userName+" about balance!");
+        this.userName = userName;
+        this.amount = money;
+    }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public Money getAmount() {
+        return amount;
+    }
+
+    @Override
+    public String toString() {
+        return null;
+    }
+}
