+++ src/testClasses/database/Registretion_Test.java
@@ -1,11 +1,5 @@
 package testClasses.database;
 
-import dataModel.models.SystemProfile;
-import database.DatabaseAuthorizationController;
-import database.DatabaseController;
-
-import java.util.Scanner;
-
 /**
  * Created by Nikolay on 12/21/2016.
  */
